requirements.txt
praw
nltk
sqlalchemy
telebot
matplotlib
pillow
